add denied access for different users

+++ backend/src/controllers/boards/delete-board.js
@@ -9,6 +9,10 @@ const deleteBoard = async (req, res) => {
             return res.status(404).json({ message: 'Board not found.' });
         }
 
+        if (board.user_id != req.user.id) {
+            return res.status(403).json({ message: 'Denied access.' });
+        }
+
         await knex('tasks').whereIn('card_id', function () {
             this.select('id').from('cards').where('board_id', id);
         }).delete();
